requirements.txt
amqp==5.1.1
antlr4-python3-runtime==4.11.1
asciimatics==1.14.0
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
binary2strings==0.1.9
bintropy==1.4.3
bleach==6.0.0
capstone==4.0.2
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama>=0.4.5
commonmark==0.9.1
contourpy==1.0.7
coverage==7.1.0
cryptography==39.0.0
cxxfilt==0.2.2
cycler==0.11.0
dacite==1.8.0
decorator==4.4.2
dill==0.3.6
Django==4.1.7
django-bootstrap-v5==1.0.11
django-extensions==3.2.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
dncil==1.0.2
dnfile==0.13.0
docutils==0.19
elflepton==2.0.2
enum-compat==0.0.3
exceptiongroup==1.1.0
executing==1.2.0
first==2.0.2
flake8==6.0.0
flare-capa==5.1.0
fonttools==4.38.0
funcy>=1.17
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
halo==0.0.31
humanize==4.5.0
ida-netnode==3.0
ida-settings==2.1.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
iniconfig==2.0.0
intervaltree==3.1.0
ipython==8.11.0
jaraco.classes==3.2.3
jedi==0.18.2
jeepney==0.8.0
jsonpath-ng==1.5.3
keyring==23.13.1
keystone-engine==0.9.2
kiwisolver==1.4.4
kombu==5.2.4
lief>=0.12.3
log-symbols==0.0.14
loguru==0.6.0
markdown-it-py==2.2.0
matplotlib==3.6.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mock==5.0.1
more-itertools==9.0.0
msgpack==1.0.4
multiprocess==0.70.14
mysqlclient==2.1.1
networkx>=2.5.1
nose==1.3.7
numpy==1.24.1
overrides==7.3.1
packaging==23.0
parso==0.8.3
pefile>=2022.5.30
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip-licenses==4.0.3
pipdeptree==2.3.3
pkginfo==1.9.6
plac==1.3.5
pluggy==1.0.0
ply==3.11
prettytable==3.6.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
pydantic>=1.10.4
pyelftools==0.29
pyfiglet==0.8.post1
pyflakes==3.0.1
Pygments==2.14.0
pygraphviz>=1.10
pykerberos==1.2.4
pyparsing==3.0.9
pyperclip==1.8.2
pytest==7.2.1
python-dateutil==2.8.2
python-flirt==0.8.6
python-registry==1.3.1
python-snappy==0.6.1
pytz==2022.7.1
PyYAML==6.0
qiling==1.4.4
questionary==1.10.0
readme-renderer==37.3
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.3.1
ruamel.yaml>=0.17.21
ruamel.yaml.clib==0.2.7
SecretStorage==3.3.3
six==1.16.0
smda==1.8.4
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spinners==0.0.24
SQLAlchemy==2.0.0
sqlparse==0.4.3
stack-data==0.6.2
tabulate==0.9.0
termcolor>=2.2.0
tomli==2.0.1
tqdm>=4.64.1
traitlets==5.9.0
twine==4.0.2
typing_extensions>=4.2.0
unicodecsv==0.14.1
unicorn==2.0.1.post1
urllib3==1.26.14
urwid==2.1.2
vine==5.0.0
viv-utils>=0.7.7
vivisect>=1.0.8
wcwidth==0.2.6
webencodings==0.5.1
yamale==4.0.4
yara-python==4.2.3
zipp==3.12.0
zope.event==4.6
zope.interface==5.5.2
flower==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-magic==0.4.27
python-tlsh==4.5.0
